refactor: Modify folder structure
