+++ frontend/js/tests/user-settings/SelectTimezone.test.tsx
@@ -6,7 +6,7 @@ import {
   textContentMatcher,
 } from "../test-utils/rtl-test-utils";
 import APIServiceClass from "../../src/utils/APIService";
-import SelectTimezone from "../../src/settings/pages/SelectTimezone";
+import SelectTimezone from "../../src/settings/select_timezone/SelectTimezone";
 
 const user_timezone = "America/New_York";
 const pg_timezones: Array<[string, string]> = [
